fix: check for socket_functions
git-svn-id: http://halodev.ontoprise.com/svn/HaloSMWExtension/trunk@11118 f80b7aa0-dfcc-4e65-a0d9-74108174cbc8

+++ deployment/tools/smwadmin/DF_Tools.php
@@ -245,7 +245,11 @@ class Tools {
 		exec("mysql --version > $nullDevice", $out, $ret);
 		$mysql_binaries = ($ret == 0);
 		if (!$mysql_binaries) return("Cannot find mysql.exe. Please include path to mysql.exe into PATH-variable.");
-
+        
+		// check if socket functions are available
+		if (!function_exists("socket_create")) {
+			return("Cannot find socket function in your PHP installation. Check if 'php_sockets'-extension is loaded in php.ini.");
+		}
 		if (Tools::isWindows()) unlink($nullDevice);
 		return true;
 	}
